Move hex vt layer generation to OutputPyramid
Unused ATM, but valuable as demo code. Didn't
want to bury it in a git commit as removed code.
